refact: moved resizing logic to a seprate dir
