+++ src/keystore/hash_table/resizing/hash_bucket_resizing_operation.h
